requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic
anyio==3.7.1
asttokens==3.0.0
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
black==23.11.0
blinker==1.9.0
cachetools==5.5.2
ccxt
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
comm==0.2.2
contourpy==1.3.2
coverage==7.9.1
cramjam==2.10.0
curl_cffi==0.11.4
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
dict==2018.11.19
dill==0.4.0
distro==1.9.0
empyrical==0.5.5
executing==2.2.0
fastapi==0.104.1
fastparquet==2024.11.0
finnhub-python==2.4.20
flake8==6.1.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
fred==3.1
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib-metadata==7.1.0
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
MetaTrader5==5.0.45 ; sys_platform == "win32"
multidict==6.6.3
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.44.0
numba==0.61.2
numpy==1.26.2
openai==1.93.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pandas-ta>=0.3.14b0
pandas_ta==0.3.14b0
parso==0.8.4
pathlib2>=2.3.7
pathspec==0.12.1
patsy==1.0.1
peewee==3.18.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
plotly==5.18.0
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
ptyprocess==0.7.0
public==2020.12.3
pure_eval==0.2.3
pyarrow==14.0.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1
pydeck==0.9.1
pyflakes==3.1.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==13.7.0
rpds-py==0.25.1
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.11.4
seaborn==0.13.2
self==2020.12.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.4
streamlit==1.28.2
streamlit-aggrid>=0.3.4
streamlit-autorefresh==1.0.1
streamlit-elements>=0.1.0
streamlit-extras>=0.3.5
structlog==23.2.0
ta==0.11.0
ta-lib==0.6.4
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
twelvedata==1.2.14
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
uvicorn[standard]==0.24.0
uvloop==0.21.0
validators==0.35.0
vectorbt==0.28.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xlsxwriter==3.1.9
yarl==1.20.1
yfinance==0.2.64
zipp==3.23.0
boto3==1.34.103
diskcache==5.6.3
APScheduler==3.10.4

pyyaml>=6.0
